Organizing things a bit. Split the two APIs up and moved authentication
